+++ src/Model/Entity/Page.php
@@ -12,6 +12,7 @@
  */
 namespace MeCms\Model\Entity;
 
+use Cake\Routing\Router;
 use MeCms\ORM\PostAndPageEntity;
 
 /**
@@ -34,5 +35,15 @@ class Page extends PostAndPageEntity
      * Virtual fields that should be exposed
      * @var array
      */
-    protected $_virtual = ['plain_text'];
+    protected $_virtual = ['plain_text', 'url'];
+
+    /**
+     * Gets the url (virtual field)
+     * @return string|null
+     * @since 2.27.2
+     */
+    protected function _getUrl()
+    {
+        return $this->has('slug') ? Router::url(['_name' => 'page', $this->get('slug')], true) : null;
+    }
 }