+++ tests/TestCase/Model/Entity/TagTest.php
@@ -20,6 +20,15 @@ use MeCms\TestSuite\EntityTestCase;
 class TagTest extends EntityTestCase
 {
     /**
+     * Test for virtual fields
+     * @test
+     */
+    public function testVirtualFields()
+    {
+        $this->assertHasVirtualField(['slug', 'url']);
+    }
+
+    /**
      * Test for fields that cannot be mass assigned using newEntity() or
      *  patchEntity()
      * @test
@@ -27,15 +36,6 @@ class TagTest extends EntityTestCase
     public function testNoAccessibleProperties()
     {
         $this->assertHasNoAccessibleProperty(['id', 'post_count', 'modified']);
-    }
-
-    /**
-     * Test for virtual fields
-     * @test
-     */
-    public function testVirtualFields()
-    {
-        $this->assertHasVirtualField('slug');
     }
 
     /**
@@ -48,7 +48,17 @@ class TagTest extends EntityTestCase
             'This is a tag' => 'this-is-a-tag',
             'MY_TAG.a!' => 'my-tag-a',
         ] as $tag => $expectedSlug) {
-            $this->assertEquals($expectedSlug, $this->Entity->set('tag', $tag)->slug);
+            $this->assertEquals($expectedSlug, $this->Entity->set('tag', $tag)->get('slug'));
         }
+    }
+
+    /**
+     * Test for `_getUrl()` method
+     * @test
+     */
+    public function testUrl()
+    {
+        $this->Entity->set('tag', 'a-tag');
+        $this->assertStringEndsWith('/posts/tag/a-tag', $this->Entity->get('url'));
     }
 }